FEAT: Adds basic shader functionality
This commit adds basic shader functionality to the engine. Currently
this does not include shader compilation, rather the shader compilation
has to have been done prior or externally. However, this is on the todo
for further shader implementation. The shader functionality is currently
implemented as a submodule that also needs to be initialized during the
render initialization.

In addition to shader functionality, this also requires some basic file
functionality, such as reading and writing, which has also been added
in this commit