FIX: Fix vulkan error for non-destroyed pipeline
This commit fixes a bug with how pipeline and pipeline layouts where
handled. Previously, the handle could get overwritten when a new surface
was added, which effectively meant that the old handles got lost. This
commit fixes this by adding the pipeline and pipeline layout handles to
the low-lever surface class instead.
This commit also includes some cleanup of code and additional error
checking during the surface-creation stages.

This commit will likely not build due to missing the include-file;
"vulkanVertex.h". This will be fixed in commits following this