REFACTOR: Change fn-signatures to bool return
This commit changes many of the functions that used to return a variable
or struct as either a copy of reference, to instead return a boolean
value describing if the expected operation succeeded or failed. The
variable is instead used in the function argument as a reference for
most cases. Also in most cases when the expected operation failed, an
error message should be output to the console, but the function also
returns a false value. Todo here is to implement error handling for the
higher levels, such that the application does not crash.

Additionally, this commit tries to expand the filling of member
variables for most vulkan structs, such that all member values contain
a valid value, either default or a specified.

Finally, many if-statements have been changed to test for constant !=
variable rather than variable != constant, to mitigate certain typoes
that instead might have made an assignment to the variable. This way
the compiler will complain rather than there being a silent error.

This commit will likely not build due to missing a include-file;
"vulkanVertex.h". There might also be other errors related to the
change of functions-signatures. This will be fixed in commits following
this