REFACTOR: Major refactor of vulkan implementation
This commit is a major re-work of the current implementation of the
vulkan renderer. This has been done, partly because I needed to learn
more about vulkan, and partly because it needed to support multiple
windows/surfaces etc.
The first part of the refactor has been to change function signatures
such that they return a boolean value for functions that can return a
boolean value. For example if a function does something that can result
in a pass/fail from the vulkan-api functions, this should determine the
outcome of that function. If a function then also sets or returns a
vulkan handle of some sort, this should be part of the function input
parameters.
The second part has been to separate the surface handling from the
device and commandpool handling, such that there is only 1 device in
use at all times, but there can be multiple surfaces/swapchains all
rendering different things. Hopefully this separation might also work
later down the line when we for example want to use a compute queue in
addition to the current graphics queue.

If compiling with this commit, the program should run and use the
juliaSet.Frag shader to output a Julia set on all four surfaces. The
julia set constant takes the current frame as input.

This has not yet been tested on the windows platform, and there is a
resize issue happening on the ubuntu-system I am currently. This
causes the program to crash when a window is resized to a smaller
window "to fast". I believe this has to do with the operating system
sending to many resize events, but there is currently no fix for this