REFACTOR: Update math-library to support constexpr
This commit adds the constexpr support to the math-library such that it
basically becomes a literal type. This allows the math-library to be
used for return types in other constexpr functions such as member-
functions in classes etc.
This commit also includes a minor restructure of the math-libraries
vector classes and an addition of a scalar constructor