REFACTOR: Rework much of the current math-library
This commit reworks most of the current math-library and adds support
for vulkan specific matrix generation. It also adds quaternions to the
library such that certain rotations can be done via quaternion-
multiplication instead of matrix multiplication.

There has been some renaming of files and all the vec-classes have been
moved into the same header-file, instead of being separate. This will
in future hopefully make cooperation between the classes easier.

Currently most of this implementation has not been tested, such that
the math operations, such as the prebuilt matrices may not be correct.
The same is true for the new implementations of for example quaternions
and the rotation via quaternions