FEAT: Add camera functionality to engine
This commit adds the starting point to a high-level camera API togheter
with a low-level implementation of a camera in vulkan. This is currently
a work in progress but should at the current state work with the
previously updated math-functionality. Along with this addition of a
camera, is also some modifications and updates to the current transform
component of the entity system.