FEAT: Add matrix inverse and determinant to matrix
This commit adds the functionality to do a matrix inverse and to
calculate the determinant of a 4x4 matrix. The matrix inverse algorithm
has been implemented based on Gauss-Jordan elimination.

This commit also adds some fixes for other math-functions such that the
c++ versions are used instead of c version (notably abs->fabs). In
addition to this, a new constant called MUGGY_EPSILON together with a
function to determine if a number should be considered zero or not.

The new matrix-functionality has only been briefly tested to verify
that values looks reasonable