TEST: Add basic math test and update renderer test
This commit adds a basic test for the matrix inverse and determinant
that was previously added. The matrix inverse outputs have been verified
with wolframAlpha, and appears to be correct.

This commit also adds some updates to the renderer test, such that the
low-level camera is called. There has however, not been any
verification of if the camera values are correct or not. This will have
to wait until we are actually rendering an item